hello-python-microservice/requirements.txt
nats-py
watchdog
asyncio